pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "v2ex-tx2json"
version = "0.1.0"
description = "Parse v2ex solana tx pages and return JSON-like dicts"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [{name = "Joe", email = "sdauwangzh@gmail.com"}]

dependencies = [
    "python-dotenv>=1.0",
]

[project.urls]
Homepage = "https://github.com/HelloWorldImJoe/v2ex_tx2json"
